fix the threadpool termination and wait-for-all logic: cv.notify_all() will ONLY notify those threads which are WAITING right at the moment when notify_all() fires! The old code caused lock-up at application termination quite often.
---

Without this, in practice it turns out sometimes a thread (or more) remains stuck for a while...

See also:
- https://en.cppreference.com/w/cpp/thread/condition_variable/notify_all
- https://stackoverflow.com/questions/38184549/not-all-threads-notified-of-condition-variable-notify-all
where one of the answers says: "Finally, `cv.notify_all()` only notified currently waiting threads. If a thread shows up later, no dice." which is corroborated by the docs above: "Unblocks all threads currently waiting for *this." and then later: "This makes it impossible for `notify_one()` to, for example, be delayed and unblock a thread that started waiting just after the call to `notify_one()` was made."

Ditto for `notify_all()` on that one, of course. ;-)

---

Other tweaks:

- when the number of tasks queued over the lifetime of the app is expected to surpass the value of INT_MAX, then have all related functions use the proper larger type: u64.
- add an optional multiplier to sleep_or_yield(), to be used when we want to increase the sleep time in some waiting / delay loops. (Not used in this code yet.)