adjust tesseract info dump code (PrintFontsTable() & PrintVariables()) to use tprintf() -> fz_error/fz_warn/fz_info APIs when compiled with MuPDF.
Also patch the codebase to *remove* the `debug_file` option parameter when compiling with MuPDF as the above APIs can be used (and are *intended* to be used) to redirect and/or duplicate info/warn/error output to file or other interfaces via registered callbacks, thus replacing and obsoleting the `debug_file` option's need.

---

TODO: the current edit/change is a bit of a quick hack in that it uses preprocessor logic plus some regrettably necessary code duplication to produce the above result. One wonders whether we should bother at all about a tesseract codebase *without* MuPDF integration, because, even when compiled as a stand-alone tool, do we intend to use the common MuPDF APIs for diagnostic and info reporting anyway, hence the question then becomes: shouldn't we just *replace* the mainline repo's code with our own variant to use tprintf() and maybe some explicit MuPDF/Fitz APIs, and then call it a day? Rather than duplicating code and fiddling with the code flow to still ensure the codebase will compile in classic sans-MuPDF build mode as well?