tesseract: add --two-pass for double analysis
This change will cause the main tesseract program to analyse every image
twice. Once with the given page segmentation mode (typically 3), and
then once with a single block page segmentation mode. The second run
runs on a modified image where any earlier blocks are turned black,
causing Tesseract to skip them for the second analysis.

Currently two pages are output for a single image, so this is clearly a
hack, but it's not as computationally intensive as running two full
runs.

(In fact, it might add as little as ~10% overhead, depending on the
input image)

WARNING: This will probably break weird non-filepath file input patterns
like "-" for stdin, or things that resolve using libcurl.