fix: finally caught the obnoxious crash at destructor/termination time, when the svmap iterator would fail internally while running in the debugger or spurious assertions about broken PIX structures would show up at application end: it turns out that std::map operator[]() ALWAYS replaces the object instance, which is bad form as we expected the *content* of the existing instance to be replaced *only*. I still don't understand *completely* what went down here, but the crux is to NOT USE: - std::map operator[] - std::map:insert() - std::map:emplace_back() where we did (the latter two having been tried as replacement for the first with *no success*), but that we MUST use std::map:insert_or_assign() instead!
A *hint* at the problem was found at https://www.geeksforgeeks.org/inserting-elements-in-stdmap-insert-emplace-and-operator/?ref=lbp where it is mentioned that operator= always replaces the existing instance. This was a bit of a surprise (as it is not mentioned at cppreference.com and elsewhere where I looked) but it gave me a new attack vector, which was to check the various insert/assign methods of std::map more closely.

WARNING: this seems to be part of a *random time compound bug* that still exists in the code, so I MAY be wrong in my conclusion here...   :-S