WIP: parameter/config handling refactoring: everything 'tesseract' (including PSM and OEM mode settings) SHOULD be configurable through the tesseract *parameters* (the '-c' variables) so that tesseract (as a library) can be invoked by any application, which can can then easily pass-through any configuration/setup parameters through that single interface.
As a consequence, to keep things simple and unambiguous, the Tesseract API Init*() function interfaces are simplified: those individual parameters disappear and are now supposed to be set through those '-c' named parameters, like everybody else.

TODO: we pass those named parameters and their values as a two string vectors, but the parameters are *typed* and I think it's smarter now to pass the named paramaters as an instantiated set that can be copied/used (ParamVector or PAramVectorSet): that way, the type checking, etc. can be done at any point in time when it's most convenient for the userland code. This "conclusion" now is due to our run-in with current tesseract config file handling, which, to my eye at least, is a tad convoluted. The new idea here is to move the config file handling to the *outside* and thus produce a fully processed set of named, valued, parameters for tesseract to peruse in the given instance.

TODO Number 2: we intend to use tesseract in a larger application where multiple simultaneous runs would be quite useful as we're running multicore CPUs and those beasts aren't maxed out, not by any measure. The usual approach is to plug in OpenMP (which works quite skitty on my rig: it does the job and then eats a metric ton of CPU while idling in the single-thread chunks of the tesseract code); our idea is to allow to run multiple OCR actions in parallel (most will be using the same language matrices anyway, so there's plenty common ground still) so that we don't have to worry about thread-safety in the tesseract codebase: what we DO have to worry about are those global parameters: those have to move into Tesseract class instance scope so that multiple simultaneous runs don't suffer from named-parameter collisions.