WIP: introducing NetworkDebugSettings, which was a bright spark for a bit today, until I realized that I need to transport a Tesseract instance pointer to some internal LSTM bits anyway, so this becomes rather ludicrous: those same fields are directly available from the tesseract instance as Parameters and when this code is invoked in tesseract tools which don't need (or want) a live Tesseract instance or two around, we can simply provide very basic default choices for those instead as they are all, AFAICT, relatively simple tools.
The way forward therefore is to revert this and plonk an OPTIONAL Tesseract instance reference in those LSTM classes for them to use at will. After all, we want Tesseract to become able to run multiple class instances simultaneously and independently (as far as that is possible) and having `ScrollViewManager::GetActiveTesseractInstance()` around as a hacky rescue for needed Tesseract instance access is not going to help in that department!