fixing the heap corruptor by inadvertently triggering a double-free: while `Image` may LOOK like it's a class that *owns* the Pix pointer, it DOES NOT: it is merely a C++ wrapper around a leptonica pointer and currently little more than that.
---

This was introduced while hunting a VERY troublesome bug that started as a *random*(!) Segmentation fault report in my batch runner, then got dialed down to a specific area in the code going berserk (image preprocessing -> LineFinder), then after a few days hunting and disabling the usual suspects (but I was barking up the wrong tree: killing all threads as this smelled like the multithreading error) it finally got reproducible today by rigorous application of the MSVC CtrDbg heap debug features built into the run-time, until the 'fixed spot' where it happens was a reported free space corruption during a `delete DENORM::rotation_`: that is not the culprit, but *somebody* just before there must have nuked some heap space. Finally it was discovered that an Image got double-freed (ColFinder::nontext_map_) which then led to the discovery that the one nuking it was GradeBlobsIntoPartitions() where inside we finally found the offending statement: `strokewidth.cpp@389:: nontext_map_ = nullptr;` which is due to the assignment operator of the `Image` class doing a destroy-of-previous-before-picking-up-the-new-value. Which SHOULD have been fine were `Image` a real leptonica Pix instance manager, but it is NOT: it is a *very rudimentary* C++ wrapper around Pix.

Of course that pixDestroy in that assignment operator didn't get in there for no reason: I do (or rather: *did*) have heap leak reports about several Image/Pix not being released at end of program, so that's what led to this bugger.

The point? preparation.h facilitates this kind of deep digging because now you can at least be sure that when you 'patch' preparation.h with, say, `crtdbg.h`, then there's no worry: everybody in tesseract will have got that one! (Ditto for compile-time settings, such as the extra define needed by MSVC to give you the M_PI constant, etc.)