Make `Image` PIX carrying class a proper C++ object-owning class instead of what was before: current codebase has another PIX/Image tesseract application crash (hard error) due to the registered source image being aliased internally (obviously, for work is done on it) and somewhere in the area of Normalization (and/or Thresholding) this image the alias gets reference-decounted (a.k.a. destroyed) down to a refcount of ZERO, thus making any PIX::data[] access illegal as you'ld be poking around in destructed/freed memory then. Which is the cause of the crash, only trouble is finding out exactly which code flow + Image instance "did it", because the Image class, as it currently stands, does not contribute positively to proper tracking of the reference counts (which are already facilitated by leptonica). Hence the decision today is to *eff that* and turn `Image` into something that behaves *properly* by reference counting and, for performance, modern C++ move semantics.
TODO: we'll have to go through the tesseract code with a fine comb after this as there's plenty spots in the code that "dealt" with the original `Image` implementation being half-baked like that by dealing with the leptonica-based refcounting by local initiative, all of which must be reviewed and fixed to correctly deal with the new approach.

TODO: part of tesseract still rides the leptonica PIX type as-is, where it would benefit code/pix management if we used a managing Image class instance instead: go through the tesseract code base and see where PIX is used directly and replace that with Image usage so we don't have to manually track and destroy images and image snippets any more.