WIP: final(?) part of code review re Image-vs-PIX usage.  To deal with the current public API which still uses leptonica `PIX*` C PIX pointers instead of `Image` class (temporary) instances, we introduce `clone2pix()` to keep the PIX refcount correct for when we are expected to behave like leptonica itself, i.e. returned `PIX*` pointers must be owned and cleared through `pixDestroy()`.
Incidentally, we pulled this baby through the debugger (again) and observed the PIX error was indeed in NormalizeImage(): see that part of the code for the fixes as this crash is now fixed by properly tracking the PIX instances (reference counts).

I still consider `operator Pix *()` to be pretty dangerous shorthand -- which could be replaced by the new `Image.ptr()` method -- as it both simplifies the look of the code when using leptonica APIs and is a danger when implicitly used in all other circumstances, such as when using our own public APIs, such as happened inside `NormalizeImage()`: `GetInputImage()` does not produce an `Image` but rather an `PIX*` and thus the refcounting went haywire in there previously. Ditto for `NormalizeImage()` itself with `PIX**` C style reference argument: another spot where the refcounting could go b0rk b0rk b0rk.